- fixed the bug of the profile dialog

+++ app/src/main/java/samurai/geeft/android/geeft/database/BaaSGetGeefterInformation.java
@@ -40,7 +40,7 @@ public class BaaSGetGeefterInformation extends AsyncTask<Void,Void,Boolean> {
     protected Boolean doInBackground(Void... arg0) {
         BaasUser currentUser = BaasUser.current();
         if(currentUser !=null) {
-            long feedback = currentUser.getScope(BaasUser.Scope.REGISTERED).get("feedback");
+            double feedback = currentUser.getScope(BaasUser.Scope.REGISTERED).get("feedback");
             long given = currentUser.getScope(BaasUser.Scope.REGISTERED).get("n_given");
             long received =currentUser.getScope(BaasUser.Scope.REGISTERED).get("n_received");
             mUserInformation[0] = feedback;
